+++ base/generate_data.py
@@ -7,6 +7,7 @@ import collections
 import os
 import random
 import config_path
+import preprocess
 
 train_path = config_path.train_path
 test_path = config_path.test_path
@@ -19,10 +20,10 @@ utils.gen_folder(dnsp_data_dir, fe_gen_dir, model_data_dir)
 
 continous_features = range(1, 5)
 categorial_features = range(5, 31)
-continous_clip = [1740, 6, 2, 2]  # 待调整
+continous_clip = [1740, 6, 2, 2]  #待调整
 
 utils.down_sampling(tr_path=train_path, label='click', outpath=dnsp_data_dir)
-utils.create_feature(data_path=dnsp_data_dir, datadir=fe_gen_dir, filename='dnsp_train.csv')
-utils.create_feature(data_path=test_path, datadir=fe_gen_dir, filename='test.csv')
+preprocess.create_feature(data_path=dnsp_data_dir, datadir=fe_gen_dir, filename='dnsp_train.csv')
+preprocess.create_feature(data_path=test_path, datadir=fe_gen_dir, filename='test.csv')
 
-utils.preprocess(datadir=fe_gen_dir, outdir=model_data_dir,continous_features=continous_features,categorial_features=categorial_features)
+preprocess.preprocess(datadir=fe_gen_dir, outdir=model_data_dir,continous_features=continous_features,categorial_features=categorial_features)
